update keychain using a fork with Windows credstore support
