+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/PuerkitoBio/goquery v1.5.0
 	github.com/aws/aws-sdk-go v1.37.8
 	github.com/briandowns/spinner v0.0.0-20190319032542-ac46072a5a91
+	github.com/challarao/keyring v0.0.0-20190721191509-6be0d53230e3
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
 	github.com/edaniels/go-saml v0.0.0-20160724042625-8c877c3ab101
 	github.com/fatih/color v1.7.0
@@ -14,6 +15,7 @@ require (
 	github.com/howeyc/gopass v0.0.0-20170109162249-bf9dde6d0d2c
 	github.com/magiconair/properties v1.8.1 // indirect
 	github.com/mattn/go-colorable v0.1.2
+	github.com/mattn/go-isatty v0.0.7 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/olekukonko/tablewriter v0.0.4
 	github.com/pelletier/go-toml v1.4.0 // indirect
@@ -22,8 +24,6 @@ require (
 	github.com/spf13/cobra v0.0.4
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/viper v1.3.2
-	github.com/stretchr/testify v1.6.1 // indirect
-	github.com/tmc/keyring v0.0.0-20171121202319-839169085ae1
 	golang.org/x/net v0.0.0-20201110031124-69a78807bb2b
 	gopkg.in/ini.v1 v1.62.0 // indirect
 )
